Can now join and leave sessions correcctly, and UI updates accordingly

+++ blueprints/jam_session/jam_sessions.py
@@ -89,7 +89,8 @@ def delete_session(session_id: int):
 @jam_sessions_bp.post('<int:session_id>/join')
 def join_session(session_id: int):
     jam_session = JamSession.query.get(session_id)
-    if not Party.query.filter_by(session_id=jam_session.id, user_id=session.get('id')).first():
+    party = Party.query.filter_by(session_id=jam_session.id, user_id=session.get('id')).first()
+    if not party:
         party = Party(jam_session.id, session.get('id'))
         db.session.add(party)
         db.session.commit()
@@ -103,6 +104,9 @@ def join_session(session_id: int):
 def leave_session(session_id: int):
     jam_session = JamSession.query.get(session_id)
     party = Party.query.filter_by(session_id=jam_session.id, user_id=session.get('id')).first()
-    db.session.delete(party)
-    db.session.commit()
-    return redirect(url_for('jam_sessions.get_sessions'))
+    if party:
+        db.session.delete(party)
+        db.session.commit()
+        return redirect(url_for('jam_sessions.get_sessions'))
+    else:
+        return redirect(url_for('jam_sessions.get_sessions'))
